Pin to zeromq versions based on the cppzmq version number.
A given version of cppzmq is compatible with the zeromq version of the
same number, but it can also be used with future minor version updates
of zeromq (it just may be that any new added API will not be supported
unless cppzmq is also updated). Therefore, this pins the host
requirement to the zeromq of the same version and the run requirement to
the same version or any minor version after it.